Task completed
Restructured test cases to be more organized.
