+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "start": "node server/bin/www",
-    "test": "mocha 'server/test' --exit",
+    "test": "mocha --recursive 'server/test' --exit",
     "seed": "node server/seed"
   },
   "dependencies": {
